backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
yfinance==0.2.51
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
openai==1.58.1
redis==5.2.1
httpx==0.28.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
pyjwt==2.11.0
stripe==14.4.1
